requirements.txt
ollama>=0.4.0
google-genai>=1.0.0
duckduckgo-search>=7.0.0
aiosqlite>=0.20.0
aiohttp>=3.9.0
rich>=13.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
streamlit>=1.40.0
